Questa bit splice undoing + Big Refactoring (#109)
* hierarchy: rename methods, allow mutable access to vars and scopes

* clippy fixes

* version bump since we are changing the API

* stream: derived signals are currently not supported

* hierarchy: store signal encoding separately from var

* fix hierarchy API uses

* clippy fix

* test: add questa sim bit slicing test

* reorganize signal module

* move `States` into `signal` module

* implement get_bit

* simplify to bit string function

* remove unnecessary pub

* hierarchy: detect adjacent ranges of the same variable

* wip: implement merging split vars

* wip: merge vars of same name

* mark derived signals and start handling them in signal source

* massive restructuring, things compile again

* pywellen fixes

* transform is used in hierarchy so it needs serde

* fixes

* remove criterion

* ci: remove bench

* update API usage

* clippy fixes

* bits should not be unioned :(

* ghw: drop bytes if necessary

* wavemem: re-add signal compression

* simplify

* hierarchy: create derived signals

* history: deduplicate and assign signal refs for derived signals

* signal source: call transform function

* writing a multi signal iterator is hard!!!

* remove all attempts at a generic multi-signal iterator

* todo: iterate

* remove old signal value implementation

* transform: wip multi signal iteration

* transform: success

* clippy fixes

* test: skip VCDs with derived signals

* ghw: fix slicing

* test: check for missing signal changes in stream

* stream: remember previous value to filter out no-op changes

* hierarchy: correctly tag derived signals

* hierarchy: opt comment

* test: better error message

* test: fix bug for signals without changes

* transform: fix assertion

* test: deal with events correctly

* stream: compute derived signals

* stream: dispatch final changes

* stream: emit fake time change when dealing with FSTs

* load_signals: restore signal size computation

* pywellen: remove type annotations since they might be stale

* pywellen: use uv

* pywellen: fix tests

* ci: bump action

+++ wellen/src/ghw/signals.rs
@@ -4,8 +4,8 @@
 // author: Kevin Laeufer <laeufer@cornell.edu>
 
 use crate::ghw::common::*;
-use crate::signals::SignalSource;
-use crate::wavemem::{Encoder, States};
+use crate::signal::{SignalSource, States};
+use crate::wavemem::Encoder;
 use crate::{Hierarchy, SignalRef, TimeTable};
 use std::io::BufRead;
 
@@ -242,7 +242,7 @@ impl VecBufferInfo {
     fn data_range(&self) -> std::ops::Range<usize> {
         // data is stored with N bits per byte depending on the states
         let start = self.data_start as usize;
-        let len = self.states.bytes_required(self.bits as usize);
+        let len = self.states.bytes_required(self.bits);
         start..(start + len)
     }
 }
@@ -269,7 +269,7 @@ impl VecBuffer {
                     signal_ref: vector.signal_ref(),
                     max_index: vector.max().index() as u32,
                 };
-                data_start += states.bytes_required(bits as usize);
+                data_start += states.bytes_required(bits);
                 bit_change_start += (bits as usize).div_ceil(8);
                 info
             })
@@ -441,10 +441,10 @@ fn set_value(bits: u32, states: States, data: &mut [u8], bit: u32, value: u8) {
 #[inline]
 fn get_data_index(bits: u32, bit: u32, states: States) -> (usize, usize) {
     debug_assert!(bit < bits);
-    let bits_in_a_byte = states.bits_in_a_byte() as u32;
+    let bits_in_a_byte = states.bits_in_a_byte();
     let bytes = bits.div_ceil(bits_in_a_byte);
     let index = bytes - 1 - (bit / bits_in_a_byte);
-    let shift = (bit % bits_in_a_byte) * states.bits() as u32;
+    let shift = (bit % bits_in_a_byte) * states.bits();
     (index as usize, shift as usize)
 }
